POC router module
