+++ backend/app/api/api_v1/routers/lookups/__init__.py
@@ -1,0 +1,15 @@
+# flake8: noqa
+from .lookups1 import (
+    lookup_geographies,
+    lookup_instruments,
+    lookup_languages,
+    lookup_sources,
+)
+from .lookups2 import (
+    lookup_document_categories,
+    lookup_document_types,
+    lookup_sectors,
+)
+from .utils import table_to_json, tree_table_to_json
+
+from .router import lookups_router
