+++ backend/app/api/api_v1/routers/lookups/router.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+lookups_router = APIRouter()
