+++ backend/app/api/api_v1/routers/lookups/utils.py
@@ -1,0 +1,42 @@
+from typing import List, Dict, Any
+
+from app.db.session import Base, SessionLocal
+
+
+def table_to_json(
+    table: Base,
+    db: SessionLocal,  # type: ignore
+) -> List[Dict]:
+    json_out = []
+
+    for row in db.query(table).all():
+        row_object = {col.name: getattr(row, col.name) for col in row.__table__.columns}
+        json_out.append(row_object)
+
+    return json_out
+
+
+def tree_table_to_json(
+    table: Base,
+    db: SessionLocal,  # type: ignore
+) -> List[Dict]:
+    json_out = []
+    child_list_map: Dict[int, Any] = {}
+
+    for row in db.query(table).all():
+        row_object = {col.name: getattr(row, col.name) for col in row.__table__.columns}
+        row_children: List[Dict[str, Any]] = []
+        child_list_map[row_object["id"]] = row_children
+
+        # No parent indicates a top level element
+        node_row_object = {"node": row_object, "children": row_children}
+        node_id = row_object["parent_id"]
+        if node_id is None:
+            json_out.append(node_row_object)
+        else:
+            append_list = child_list_map.get(node_id)
+            if append_list is None:
+                raise RuntimeError(f"Could not locate parent node with id {node_id}")
+            append_list.append(node_row_object)
+
+    return json_out
